Renamed customer_logging --> logging
