+++ u8timeseries/metrics/metrics.py
@@ -6,7 +6,7 @@ List of metrics
 import numpy as np
 from typing import Tuple
 from u8timeseries.timeseries import TimeSeries
-from ..custom_logging import raise_if_not, raise_log, get_logger
+from ..logging import raise_if_not, raise_log, get_logger
 from warnings import warn
 
 logger = get_logger(__name__)
