+++ u8timeseries/models/baselines.py
@@ -5,7 +5,7 @@ Implementation of baseline model.
 
 from .forecasting_model import ForecastingModel
 from ..timeseries import TimeSeries
-from ..custom_logging import raise_if_not, time_log, get_logger
+from ..logging import raise_if_not, time_log, get_logger
 import numpy as np
 
 logger = get_logger(__name__)
