+++ u8timeseries/models/regressive_model.py
@@ -5,7 +5,7 @@ Implementation of an Regressive model.
 
 from abc import ABC, abstractmethod
 from ..timeseries import TimeSeries
-from ..custom_logging import raise_if_not, get_logger, raise_log
+from ..logging import raise_if_not, get_logger, raise_log
 from typing import List
 
 logger = get_logger(__name__)
