+++ u8timeseries/models/theta.py
@@ -5,7 +5,7 @@ Implementation of an Theta model.
 
 import statsmodels.tsa.holtwinters as hw
 from ..timeseries import TimeSeries
-from ..custom_logging import raise_log, time_log, get_logger
+from ..logging import raise_log, time_log, get_logger
 from .forecasting_model import ForecastingModel
 import numpy as np
 import math
