+++ u8timeseries/tests/test_logging.py
@@ -7,7 +7,7 @@ from testfixtures import LogCapture
 from ..timeseries import TimeSeries
 from u8timeseries.utils.timeseries_generation import linear_timeseries, constant_timeseries
 from u8timeseries.models.theta import Theta
-from ..custom_logging import raise_log, raise_if_not, time_log, get_logger
+from ..logging import raise_log, raise_if_not, time_log, get_logger
 
 
 class LoggingTestCase(unittest.TestCase):
