+++ u8timeseries/utils/timeseries_generation.py
@@ -7,7 +7,7 @@ from pandas.tseries.holiday import USFederalHolidayCalendar
 from typing import Union
 
 from u8timeseries.timeseries import TimeSeries
-from ..custom_logging import raise_if_not, get_logger
+from ..logging import raise_if_not, get_logger
 
 logger = get_logger(__name__)
